Added assert match operator.
